Fixed several issues:
(1) We no longer mock the rsync.  This is because the social agreement to keep the code in sync was not maintained (possibly because the test version of the rsync code was the only portion that had a comment regarding maintaining the duplicate code).  The PSE now detects if it is in a test mode and takes appropriate alternative action.  This means the code is single-sourced.

(2) We no longer pass a magic flag into the PSE::RunStarted->commit() method.  The detection of being in a test state is sufficient.

(3) Windows path is fixed.

(4) Removed warnings due to undefined parameters, depending upon the tests run.

(5) Updated the rsync disk group.

(6) Updated percent_capacity to be 5% again, now that I understand the assumptions made.  This code still needs to be further updated.

(7) Updated the resource locking mechanism to delete the lock in schedule and confirm.

(8) Moved rsync disk parameter to add_param_run_directory, since parameters added after confirm do not actually get put into the database.



git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@26302 127c0c0e-971a-0410-ad48-868c1a942d8d