Spit out en error message if $Config{'use64bitint'} is false
Add POD for all (most?) methods

Remove all references to chromosome lengths

The token for the end of the linked list is now null (0) and
not 2^64-1

The old merge() method is now to be used only for unsorted
source files.  A new method, merge_sorted() is for sorted files
Under the hood, it's using get_alignments_for_position() vs.
get_alignments_for_sorted_position()

New method get_alignments_for_sorted_position() does some
calculating for which records are all part of the same
alignment slot, and does a single read() to get the 
data, instead of walking the linked list



git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@26377 127c0c0e-971a-0410-ad48-868c1a942d8d