Refactored the file naming logic into Genome::Model::Event.
screen-reads eliminate-all-duplicates nominally works

align-reads maq is now responsible for making the .bfq file instead of accept-reads
The .bfq is made from either the unique or original fastq file depending on whether we
screened the reads

assign-run solexa doesn't make the .dbm index file anymore.  It was too slow

The top level add-reads step can now take a run-name or a full-path.  It tries to
figure out what lanes to work on automaticaly - checks the DNA in that lane with the
model's sample name


git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@32698 127c0c0e-971a-0410-ad48-868c1a942d8d