align-reads maq now outputs a file of unaligned and low-quality reads
process-low-quality-alignments maq takes the file and creates a fastq from it

The top-level add-reads now uses 2 methods to determine the full path to the gerald
data.  The new one uses the "correct" method to find a "configure alignments" step and
look up a gerald_directory pse param.  The old method just find the top-level data directory
through GSC::Equipment::Solexa::Run, and glob()s to determine a gerald dir.  If the
first method fails to find any info, it tries the second




git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@32753 127c0c0e-971a-0410-ad48-868c1a942d8d