Changed the event processing flow so that the top level add-reads steps pre-creates the event/command objects in the 'Scheduled' state, and when they get executed, it will re-load that event object instead of creating a new one
Genome::Model::Event now implements execute().  And its just a wrapper
around all the lowest-level command object's _execute().  This makes it
possible for the event rows in the DB to get date_completed and event_status
attributes filled in without the lowest-level commands doing it by hand.
When we create a wrapper command to load and execute these things on the
blades, this will have to at least be partially backed out.


git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@32837 127c0c0e-971a-0410-ad48-868c1a942d8d