Renamed execute() to _execute()
Got rid of the code to iterate over lanes, as each instance only
handles one lane now

Catch the stdout and stderr from running maq map.  If it had a
problem, print out the error.  If it worked, save the line that
contains the string "match_data2mapping" to a file with the
extention .matchdata



git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@32839 127c0c0e-971a-0410-ad48-868c1a942d8d