Broke out the logic to pass control to the genome modelling pipeline into a separate method.
Removed old code which presumed one step would be called for multiple libraries.  This step is now single run/library, and operates on all lanes for a run for a given library.

Added code to force through non-deduplicated reads, which is disabled but can be enabled by developers.

Cosmetic cleanup.
 

git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@34008 127c0c0e-971a-0410-ad48-868c1a942d8d