We now schedule all the ImportSequences.
related_seq_id is now just genome_id.  We let the ImportSequence worry about
what chromosome is relevant to the situation.  This provides the advantage of
being able to decouple the random chromosomes from the rest nicely and give
us the power to schedule all the imports at once.

Improved a few error messages to help with my debugging on the way.

Ensured that the test confirms regular fastas first and then random fastas
afterward.  This allows us to have the chromosomes in place before their
related random pieces.


git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@36715 127c0c0e-971a-0410-ad48-868c1a942d8d