_loading_was_done_before_with_a_superset_of_this_params_hashref() does a more thourough check by going though combinations of input params, and not just one at a time
_create_import_iterator_for_underlying_context()'s closure handles some subclasses
differently.  If an object is loaded that requires subclassing, and that subclass
requires a new query to include an additional join not present in the original query,
it now tracks those sub-iterators by the most specific subclass with a table, not just
the subclass.  Some sub-sub-classes may not have a table when the subclass does.


git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@39911 127c0c0e-971a-0410-ad48-868c1a942d8d