Removed code duplication.
This was part of a demonstration of techniques for identifying and removing
duplication.  In this case, I chose to make a local subroutine.  Note that I
had to use fun func_name=sub syntax, since the intent here was to close over
known variables.  Within this method's context, it seemed silly to accept the
duplication of the same five other variables.

git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@41622 127c0c0e-971a-0410-ad48-868c1a942d8d