Fixes a bug that occurs when you pass in a listref of values for one of the required_for_get properties in a FileMux datasource.
It used to use the constant_values hashref of the loading template to 
pre-fill-in the values that are part of the class but don't exist in 
the data source.  The new method alters the loading template to
include these constant values (so the object fabricator knows 
that they'll be in the data returned by the data source) and pushes their
values onto the end of the data returned by the data source.

git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@44311 127c0c0e-971a-0410-ad48-868c1a942d8d