For 'via' properties, you can now use "to => '-filter'" to indicate that this property is a subset/filter of another is_many property.  For example:     things => { is => 'URT::Thing', reverse_id_by => 'me' },     white_thing => { via => 'things', to => '-filter', where => [color => 'white'] }
a property with 'is_class_wide => 1' now creates accessors with mk_rw_class_accessor and
mk_ro_class_accessor.  The accessors use a variable trapped in a closure, so all calls
to the accessor method use the same scalar for storage.

Better error messages for indirect/via/object properties when there's problems with the
class/property descriptions, and it can't figure out where the data lives.


git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@46569 127c0c0e-971a-0410-ad48-868c1a942d8d