Fixed a bug where the order of items returned from a get() with -order could be different whether the items are retrieved from the DB or the object cache.  The sorter for object cache items included 'id' as the last sort property, but the generated SQL did not include id columns.
The fix involved having the class data and loading templates keep track
of the order by columns as a listref instead of the order by clause as 
a string.  Then, the user's order by list can be prepended to the front
of the list when necessary.  Finally, the SQL order by clause is constructed
at the last moment.


git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@49426 127c0c0e-971a-0410-ad48-868c1a942d8d