Switch to using C++ version of breakdancer and account for backwards compatibility with the executable name and path change.
remove breakdancer_params and bam2cfg_params. Stick with sv_params (colon separated) which is what the processing profiles use. This will not be a tool sometime soon anyways, and just a build.

sv detector and params are no longer optional, and probably should never have been (params can still be effectively blank)

removed version hack in create that was setting the version to default if it was blank. All somatic processing profiles have sv_detector_version specified now so this is not an issue anymore.