Move tumor/normal model resolution for Somatic models into Genome::Model::Somatic->create();
Resolution of the input models for Somatic models is now handled within create();
The type_specific_parameters_for_create() added for ref. align. has been used to
suport this.  The hack in `genome model copy` has been removed in favour of explicitly
adding tumor_model_id and normal_model_id as parameters to define and as properties
of Somatic models.