Explicitly define Genome::Model::Build::SomaticCapture for compatibility with FileSummary report.
The SomaticCapture pipeline still relies on several components of the main Somatic pipeline,
including `gmt somatic run-reports` for report generation.  The FileSummary report requires
certain functions to be implemented on the build object.  Although there is a large set of
files that are only produced by one or the other pipeline, those which are produced by both
are now reported properly for SomaticCapture builds.