Create versions of Somatic reports tailored to the SomaticCapture pipeline.
The SomaticCapture pipeline currently uses the existing Somatic pipeline's
RunReports tool, but that tool assumes it is working with a Somatic build
and so looks for some things that aren't present and ignores others that are.

The FileSummary report now looks for the correct inputs for the SomaticCapture
pipeline.  The Variant report is identical to the Somatic pipeline, but the
(heretofore unused) code in G:M:SomaticCapture::Report::Variant (which was a
copy+paste of an older version of the Somatic report) has been updated to just
refer to the Somatic version.

This commit contains the new `genome model somatic-capture run-reports` and
corresponding FileSummary and Variant reports that it runs.  This RunReports
module will be added to the pipeline in a later commit.