First version of a command to generate a decision tree based on transcript annotation information.
A new command, `genome model convergence find-significant-attributes`, that takes the reference
alignment models in a convergence model and uses Weka to generate a decision tree for predicting
"tumor" or "normal" based on the presence of variations within individual transcripts.

This version contains a few hacks/less-than-ideal methods of getting things... but it works on
at least one set of 100 models!  Notably there is code borrowed from UR::Context in order to
make the iterator consistent.