Use G:ID::FlowCell->lanes() instead of `sqlrun`. Support --force.
There is now a Genome::InstrumentData::FlowCell object to get the lanes,
so we don't need to rely on a manual query to solexa_lane_summary (which
is not even the source of our data any longer).  The --force option now
works for a --flow-cell-id as well in `genome model instrument-data assign`.

The testcase has been updated to not mock the lanes due to a UR conflict
between the mocked lanes and the lanes loaded by G:I::FlowCell->lanes.