Blithely updated expected results to match updated VarScan.
`gmt varscan germline` depends on the version of VarScan located in ~dkoboldt,
which was updated recently.  This update to the test points to a new expected
result, which accounts for these changes.  The only modifications are in some
of the statistics &c. for reported snps (snvs)--however which snps (snvs) are
called remains unchanged (as evidenced by the identical .bed files in v3 and
v4 of the expected results.