Somatic builds now run their workflows directly.
When the Somatic pipeline was first created, it used a staged
processing profile with one stage with one event that called a tool
that ran a workflow.  Now that builds can grab workflows to run
directly, all of that indirection has been removed.  A side-effect
of this change is that Somatic workflows can no longer be run outside
of builds; however the logs are now stored with the build and
the build process is much simpler.