Restore feature of looking up workflow instances without instantiating them.
Old workflows cannot even be instantiated if one of the operations that was
run cannot be found (e.g. Genome::Model::Tools::Somatic::Sniper was moved).
Try to do things the quick and object-layery way, but fall back on direct
SQL by name.  New test case that includes one build for each of the two
cases (at present--if the operations change, the later build would need to
be updated).