Work toward speeding up the annotation process.
The major change is that it now intersects variation position data directly
with TranscriptStructures, instead of going through Transcript information
first as before.  The VariantAnnotator no longer deals with Transcript
windows.

Created a class/datasource called Genome::TranscriptStructure that is a
duplicate of the existing Genome::TranscriptSubStrcuture.  The new class
stores all the structures in the same chromosome in one file.  This is so
older builds can still be re-run with older snapshots.  When the older
data is no longer needed, we can remove the older classes.

Removed lots of duplicated code.