Put .bases files into a subdirectory in ImportedReferenceSequence builds.
The reference sequence might contain tens of thousands of contigs that
each get a .bases file made for them.  This makes examining the build
directory a cumbersome task, so move them out of the way.

This commit contains code for compatibility with both possible locations.
Once all existing IRS builds are updated it can be removed.