Refactor model-group coverage report to work for arbitrary model sets.
The main logic has been moved into the set view and the XSL has been moved
to be on the Genome::Model class itself so it is accessible from both views.
This step is in preparation for adding links to coverage views of model sets
on certain other status pages (e.g. samples).