Store CoverageStats logs in the build directory for Ref. Align. builds.
Add an optional parameter to `gmt bio-samtools coverage-stats` to direct
the logs generated by the workflow to a specified location. Use this
parameter in the pipeline to store the logs in the build's log directory.

(Still falls back on the hard-coded path in /gsc/var/log/ if a path is not
provided.)