TranscriptVariants writes its output to a tempfile and mv's it to --output-file
TranscriptVariants.pm now writes its output to a tempfile and then uses a mv to
move its tempfile to the place in --output-file.  This occurs in all cases
except when the output-file is specified as STDOUT (the default).  This should
resolve some of the skip if output present issues in annotation with the somatic
pipeline.