Refactored genome model build commands to use a common Base
Since most of the build commands need to limit to builds run by the user I
created the Base.pm class and refactored the commands to use it. Now all
the commands that actually work on builds limit to run_by=user. Remove
also limits to error-free builds and instrument data in preparation for
deprecating remove-abandoned.