Changed annotation cache creation to mirror updates
Previously, annotation cache creation relied on the use of an _updating
directory as an implicit lock for cache creation.  This contrasted
updating's apprach, which used a Genome::Utility::FileSystem lock.

Annotation cache creation now uses a similar locking mechanism as
updating.  Refactors have also been made so that this code is not
repeated everywhere.