Initial update to gmt graph mutation-diagram to use new interpro results
Initial update to gmt graph mutation-diagram to use the Genome::InterproResult
objects from the annotator.  As a result, added a dependancy of models and
builds.  As a part of this, most dependencies on the MG namespace were removed.

The modules in GMT::Graph::MutationDiagram::Mutation::Diagram may
still depend on the MG namespace.  This is a target for a future release.

The structure is still really ugly.  MutationDiagram/MutationDiagram.pm
isn't even a UR object.  This will be refactored in later commits.