Even more annotator speed improvements
Override the data source loading iterator for Genome::DataSource::TranscriptStructures
to allow the annotator code to filter data coming in from the file and only pass
structures that intersect up to the object fabricators and context tracking

This version generates output consistent with the currently-used/old annotator code and
is significantly faster.  Old annotator finishes a 58k variation list in 3h 15m, this
new code finishes in 10 min.