A different bug-compatability "fix" with the way it annotates deletions
For deletions, only do the regular annotation for the first structure of each
transcript.  Keep a note of the other structures and put their info in the special
deletion-data column of the output.

NOTE that this really needs to go away when the users are comfortable with doing
proper annotation of deletions.  The old annotator did this out of necessity
because of the way the data was structured.  The new code does not have this
limitation.