Only keep the first structure in a transcript for further annotation
This is the next part in a continuing series on bug compatability hacks to
mimic the output of the old annotator code.

NOTE that there is at least one case where this will produce demonstratabily
wrong answers.  If we're running with the "top" filter (which the pipeline
does at the moment), the variant is an insertion, the insertion point sits
on the boundary between two structures that butt up against each other, and
the second structure is more "important" than the first, then it will
incorrectly report the annotation on the first (less important) change.  The
old annotator code also has this bug.