Patch for bug compatability in the old annotator code
Keep only those substructures where the variant's start position is
between the endpoints of the substructures.

The windowing_iterator considers the start and stop positions for both the
variant and the substructure, but the old code doesn't.  This actually shows
up as a bug in the old code where a deletion occurs just before a transcript,
and extends past the starting point of the transcript.  The old code would
not find any intersecting transcripts/structures because it only looked at
the start position of the varaint, which does not fall within the transcript's
range.  The new windowing_iterator, doing a proper intersection, does find an
affected substructure.