Allow creation of capture container (in LIMS) at feature-list creation time.
Just calls the script that LIMS has for creating capture sets if the option
is specified; the capture container is not in any way subsequently linked
to the feature list (except that their names are the same). This will result
in two copies of the BED file, one in each system.