Updated gmt graph mutation-diagram to use interpro data from annotation data
gmt graph mutation diagram used to grab its protein domain data from the old
data base table that was accurate at around ensembl 48 (or so).  This has been
changed to hit the annotation data interpro results (the current standard for
protein domain data).  As a result of this, the tool now requires a --reference-
transcripts parameter similar to the one used in annotation.

This release also removes the --maf option, which allowed maf files to be used
as input in addition to annotation data.  This was made at the request of MG.

This also removes the use of the MG namespace from this tool.  The structure of
this tool is still awful, and many of the modules aren't UR objects.  That
is a project for another day though...