Fix for the _apply_indel_and_translate() method when it gets all the structures for a transcript
Earlier, I added a hook for the TranscriptStructure data loading so it would
only create objects for structures we were interested in.

Unfortunately _apply_indel_and_translate() needs to be able to get all the
structures in a transcript without that extra filtering.  localize the filtering
hook during the get() and it'll get all the structures it's interested in