Remove deprecated reference_sequence_name parameter from RnaSeq profiles.
Now that no running builds rely upon this value and the inputs exist, the
values have been removed from the database and this commit removes the
parameter from the processing profile.  That should complete the transition
for reference sequences to inputs.  (Annotation remains a parameter as it
does currently in ReferenceAlignment.)