Tolerate incomplete PSEs--just do secondary assignment and mark them.
A parameter is logged on the PSEs that model generation was not
attempted due to incomplete parameter specification.  However, the
case of no processing-profile-id specified is now tolerated.  (It has
been normal practice in QIDFGM for some time, but we had been holding
them back.)  If our processing standards change, the previously skipped
data can be reached by the parameter added.