Model inputs that are models are copied to a build as the last_complete_build.
CHANGELOG Any input assigned to a model that is itself a model will be copied
to the build as its last_complete_build.  If the model used as an input has
no applicable build then creating a build for the model to which the input is
assigned will fail.

This is the first step in the process of converting model and build links to
inputs.  Once this code is in the stable snapshot then we can start creating
inputs to mirror the links.