Partial fix of a bug where deletions get annotated improperly
If a deletion goes past the end of the structure, the old code truncated the
variant down to fit.  And if the deleted portion included a splice site,
then it could improperly get reported as a frame shift deletion instead of
a splice site deletion.

It still doesn't get the right answer in the default case, but now it will
when the --get-frame-shift-sequence flag is turned on