Added manifest to imported reference sequence builds and support stuff
CHANGELOG: ImportedReferenceSequence builds now get a manifest file containing
the path, size, and md5 of each .bases file in the build.  The manifest is
created when the build is, and manifests have been backfilled for each
well-formed imported reference sequence build*.  The manifest file path
can be obtained from the build via the manifest_file_path property.

*Well formed implies that there are .bases files in the data directory or
in the bases subdirectory.