Moved Bed converter to the annotator version directories and use them
Added the BedToAnnotation converter to the annotator's version directories and
converted the call in transcript variants to use the correct version converter.
Ref. Align and other pipelines have a step that calls the converter, which
has not been changed.  This will only really affect PINDEL.

If Chris' bed insertion format is adopted, these changes will need to be pushed
up to everything else that used BED, including the converter in
GMT::Bed::Convert and the variant detection stuff.