Support processing instrument data in formats other than 'Illumina' in RnaSeq.
The FASTQs in the build directory will match the format of the original
FASTQs, and the appropriate options will be set in the FASTQ-to-SAM
conversion and in Tophat itself.

CHANGELOG: The RnaSeq pipeline will try to work on FASTQs with quality scores
in Illumina, Solexa, or Sanger format instead of assuming Illumina format.