Changed how model inputs for models get copied to build inputs for builds.
Instead of always using last_succeeded_build, now we check if the build
already has a value for the input (via create params), and if not, use
the value returned by G:M:Build->_select_build_from_model_input (which
can return undef to force the create param to be required).

G:M:ImportedAnnotation and its build object are now set up to use this
for their reference_sequence.