Versioning for Varscan. (Now other tools can use GMT::Varscan to get a version)
There is a special version, 'latest', that will point to a symlink to the
latest version instead of using a specific one.  This is different than the
'default' version that is specified in the module.  In the future, all pipeline
code should request a specific version rather than relying upon either of these
options.