Further simplify the looking up of compatible instrument data.
Don't bother going through the sample name since that is derived from
the sample in all cases.

CHANGELOG: Compatible instrument data found for `genome model instrument-data
assign --all` for a model of a taxon now includes any sample that either links
directly to the appropriate taxon_id or to an individual that links to that
taxon_id.