Remove obsolete mechaism for getting ROI list.
The only models with the processing profile parameter defined but no
ROI list pre-date coverage stats and thus their current approach to
analysis is preserved.

Set is_many and is_mutable on the region_of_interest_set_value so
that `genome model input remove` works.  (This probably shouldn't be
an is_many property, really... but this gets it to a more usable
state than before when none of the genome model input commands worked
on the property.)