CHANGELOG: dbSNP concordance bugfixes, reports will change
previously, quality levels for which there were called snvs but no hits
in dbsnp were not included in the concordance calculation. fixing this
potentially lowers concordance scores.

additionally, there was an array indexing problem causing some data to
be excluded from or presented at the wrong quality level in the graph.