CHANGELOG: renamed Genome::Utility::FileSystem to Genome::Sys
The actual code is in Genome::Site::WUGC::SysDeprecated, though it
puts functions in the Genome::Sys namespace.

An empty Genome::Sys represents the portion of that namespace which is
in use by released modules.  As released products use parts of that API,
we will polish/rename as needed, and move them from the deprecated to
the official module.