Make `genome model somatic-validation define-models` require fewer arguments.
A capture set and a somatic model are all that is needed to create all the
validation models.  A feature list will need to already exist in the system
associated with the somatic model.

This current commit still requires the processing profiles to be specified as
no defaults are filled in, even though they are listed in the optional section.