always allow reallocate when we are reporting the true size (via du)
CHANGELOG: Always allow reallocate when we are reporting the true size
via 'du'. Since 'du' is reporting the actual size we will allow it to
over-allocate for the sake of having a true state in the allocation
system. If things are triggering over-allocations then the initial
estimate needs to be improved so that it does not cause
over-allocations.