Added code to allow linking reference sequences together
This can be used to creates a single logical reference sequence out of
several existing ones. The property controlling this is append_to.

e.g., if $ref1 and $ref2 are ReferenceSequence objects,

$ref2->append_to($ref1)

means $ref2 now consists of all of $ref1, followed by all of $ref2.

AlignerIndex understands this and will do the right thing when these
types of references are encountered.

The API is not exposed to genome commands yet, as it is still in
testing.