Added ei_id to the AQID test to fix seg faults
LIMS is trying to make ei_id non-nullable on all PSEs.  As part of this,
Paul added a monitor to GSC::PSE to track the creation of PSEs without ei_ids.
Something here broke, causing a bunch of seg faults with our no-commit PSEs.
Added (my own) ei_ids to our no-commit PSEs, as this was going to break sooner
or later.  This ei_id should probably be changed in the future to a better
dummy value, but it'll work for the time being.