New "build_needed" flag, the calculated companion to "build_requested".
Should be 1 when there is no complete/running build or the model has
changed since the last build was built.  The basic implementation
does this by comparing all the links and inputs and ensuring they
point to the same thing.

Convergence models already had an implementation of deciding if they
need a build; just use it instead.  ReferenceAlignment and SomaticVariation
have differences where the build input is a build and the model input is
a model; this could possibly be generalized.  ReferenceSequence models
never need rebuilding.