Gatk SomaticIndel caller will now ignore extra sequences
where extra means sequences not found in the reference fasta it is
using. the motivation for this comes from the shady business of
detecting variants in imported bams aligned against foriegn references
with /mostly/ the same sequences as ours, but a few extra.

the way this is implemented, if /a/b/c.fa is the path to the reference
sequence, then if there exists a file /a/b/seqdict/seqdict.sam, any
sequences listed in that sam file will be "whitelisted", and anything
else ignored.