Tophat results now produce an unaligned reads BAM in lieu of a combined BAM.
The tophat-alignment-stats tool will now produce a BAM with the reads Tophat
did not align instead of a combined BAM with the Tophat alignments included.
We keep the original Tophat accepted_hits.bam so this change eliminates
redundant information and the associated extra work to sort, index, and
flagstat the giant BAM.

CHANGELOG: Tophat alignment results (such as in the RNA-Seq pipeline) no
longer produce an all_reads_merged.bam file. In addition to the standard
accepted_hits.bam output from Tophat will be an unaligned_reads.bam containing
the reads Tophat did not map.