Tool to copy the 58_37c_v2 RibosomalGeneNames.txt file into annotation builds
This tool copies the 58_37c_v2 RibosomalGeneNames.txt file into Ensembl
annotation builds.  This file is a pre-req to create all of Jason's files for
RNA seq. The file was hand curated by tech-d, there is no automated means to
generate versions for different annotation data sets.  This appears to be the
least evil option.