Updates to music play: bring back mutation-relation, add missing options.
Several options were added to other commands and these needed to be added
as options in music as well in order to get them to succeed.

Now that mutation-relation will accept the results from smg to narrow its
scope, bring it back into the fold.  It will be run last in the sequence.