IntermediateAlignmentResults can now shortcut and be used on BWA re-runs.
Update the munging of the input_file input so that get() can find an
object create() made using the same parameters.

Run the BWA IntermediateAlignmentResult in a separate process so it can
commit if successful even if the final result subsequently fails.

CHANGELOG: BWA alignment will now shortcut `bwa aln` on a re-run if later
steps in the alignment process fail.