A way to inject a static HTML snippet into the body of a view.
It functions by sticking the snippet into the middle of an XML document
so the HTML needs to be XHTML (i.e. valid XML).  The standard XSLT is then
applied but with an extra stylesheet that extracts the HTML document in the
middle of the chrome.

[For a less complicated means of achieving static HTML, one can create the
relevant XSL and ignore the XML document autogenerated by the view.  However,
the method presented here allows it to style itself in the fashion of another
perspective as desired.]