rewrite to remove create_or_update
Since create_or_update was making a get call it was holding on to a
database handle for an unneccesarily long time. This was re-written to
not use a get call and instead rely on the items being ID'd by UUIDs and
will have the IndexQueue daemon infer whether it should add or delete an
item.