modified observers for project and model group
the delete observers for these classes required a get to determine if a
delete should or should not be done. for example, the model group
pending delete will have already been loaded, but the query done by the
project observer would be slightly different (asking for uuid instead of
id). UR would see the already loaded model group had been changed
(deleted), which doesn't jive with what the database returned, and throw
an exception. this is now fixed by removing the need for another get
through a slight refactoring of the observers.