Fix a bug in the processing of genotype files
The code was written to process two files and match the data where the
chromosomes and positions were the same.  In practice, it was only ever
used by passing in the same pathname for both files.  The code was
simplified to only operate on one file

In addition, there was a bug where if the first file had switched chromosomes,
and the previous line from the second file had a position numerically lower
than the new line from the first file, it would not read from the second file.
This simplification fix that problem, too