DV2 using alignment result IDs until objects are sorted out
DV2 was attempting to use the actual alignment result objects instead of
their IDs, but this was leading to problems with polymutt where a single
result was generated for a particular set of alignment results but then
that same result was used to shortcut all subsequent polymutt runs. This
has been temporarily fixed by switching to using IDs instead.