Tool to create a manual build from DBSnp web data.
Tool that will fetch a VCF file from the DBSnp FTP site, along with the
corresponding flat files and merge submitter data from the flat files
into the VCF file. This is combined with a command to import the created
VCF file into a manual result which is used to define an imported
variation list model.