Allow captures when finding matching files to diff to narrow the search space.
Previously in regex_files_for_diff one could write something like:

directory/subdirectory\d/file\d.ext

if a digit would be added to the file path but we wanted to compare the two
files named "file#.ext" regardless of the numbers in the path. However, if
there were multiple "subdirectory#" entries they would have to be explicitly
listed to avoid matching too many in the second build:

directory/subdirectory1/file\d.ext
directory/subdirectory2/file\d.ext

Now this can be simplified by using a capture in the regex:

directory/subdirectory(\d)/file\d.ext

When a file matches this regex in the first build, the captured value is
substituted in place of the capture so tht only a "file#.ext" in an
identically named subdirectory will be found for comparison, even if that
subdirectory name is not known in advance.