Remove Default for import-interpro's tmp directory and make it non-optional
gmt annotate import-interpro wouold dump all of its (non temporary) temp files
to $self->tmp_dir.  By default, this was optional and would be set to /tmp.
This is all kinds of awful for reasons including requiring all jobs to run on
the same host by default (making this super slow).

This commit alleviates some of the shame that should have been festering for
two years...